+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- improve "missing a main file" error when adding multiple components to print the problematic components
+
 ## [13.0.5-dev.16 - 2018-10-14]
 
 - improve `bit export` performance by avoid loading components from the filesystem to generate links